Refactor recommendation generation
Moved all of the logic for retrieving a recommendation from the Places API into app.route("/recommend"), which is called when a button called "Find a Place" is clicked. This button is only generated after the weather is retrieved. This button disappears after it is clicked. A similar button called "Find Another Place" executes the same logic, but only appears after a recommendation has been created.